+++ modules-template/prov-template-compiler/src/main/java/module-info.java
@@ -22,6 +22,7 @@ module prov.template.compiler {
     exports org.openprovenance.prov.template.compiler;
     exports org.openprovenance.prov.template.compiler.past;
     exports org.openprovenance.prov.template.compiler.past.annotations;
+    exports org.openprovenance.prov.template.compiler.past.checker;
     exports org.openprovenance.prov.template.compiler.past.emitter;
     exports org.openprovenance.prov.template.compiler.past.type;
     exports org.openprovenance.prov.template.compiler.configuration;