+++ modules-template/prov-template-library/src/main/python/past/util/List.py
@@ -1,6 +1,6 @@
 
 
-class List:
+class List(list):
     def __init__(self):
         self._list = []